fix: moving restore logic to backend for security reasons
