+++ components/utility/profile-tabs/subscription-tab.tsx
@@ -1,16 +1,15 @@
-import { FC, useContext, useState } from "react"
-import { useRouter } from "next/navigation"
 import { Button } from "@/components/ui/button"
-import { Label } from "@/components/ui/label"
 import { Input } from "@/components/ui/input"
+import { Label } from "@/components/ui/label"
 import { TabsContent } from "@/components/ui/tabs"
 import { PentestGPTContext } from "@/context/context"
 import { getBillingPortalUrl } from "@/lib/server/stripe-url"
-import { restoreSubscription } from "@/lib/server/restore"
-import { toast } from "sonner"
+import { SubscriptionStatus } from "@/types/chat"
 import * as Sentry from "@sentry/nextjs"
 import { IconRefresh } from "@tabler/icons-react"
-import { SubscriptionStatus } from "@/types/chat"
+import { useRouter } from "next/navigation"
+import { FC, useContext, useState } from "react"
+import { toast } from "sonner"
 
 interface SubscriptionTabProps {
   value: string
@@ -47,21 +46,34 @@ export const SubscriptionTab: FC<SubscriptionTabProps> = ({
   const handleRestoreButtonClick = async () => {
     try {
       setLoading(true)
-      const restoreResult = await restoreSubscription()
-      if (restoreResult.type === "error") {
-        Sentry.withScope(scope => {
-          scope.setExtra("user.id", profile?.user_id)
-          Sentry.captureMessage(restoreResult.error.message)
-        })
-        toast.error(restoreResult.error.message)
-      } else {
-        if (restoreResult.value === null) {
-          toast.warning("You have no subscription to restore.")
-        } else {
-          toast.success("Your subscription has been restored.")
-          updateSubscription(restoreResult.value)
+      const response = await fetch("/api/stripe/restore", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json"
         }
+      })
+
+      const data = await response.json()
+
+      if (!response.ok) {
+        throw new Error(
+          data.error || "An error occurred while restoring the subscription"
+        )
       }
+
+      if (data.message) {
+        toast.warning(data.message)
+      } else if (data.subscription) {
+        toast.success("Your subscription has been restored.")
+        updateSubscription(data.subscription)
+      }
+    } catch (error: any) {
+      console.error("Error restoring subscription:", error)
+      Sentry.withScope(scope => {
+        scope.setExtra("user.id", profile?.user_id)
+        Sentry.captureMessage(error.message)
+      })
+      toast.error(error.message)
     } finally {
       setLoading(false)
     }
